requirements.txt
mistralai
chromadb
streamlit
beautifulsoup4
python-dotenv
requests